修正排版:去掉多餘空白與型態轉換
delete spaces and data type converting on line 16

+++ board/src/MessageRepository.php
@@ -13,7 +13,7 @@ class MessageRepository extends EntityRepository
     {
         $dql = "SELECT count(m.sn) FROM Message m";
         $query = $this->_em->createQuery($dql)->getSingleResult();
-        $total = (int) ($query[1]);
+        $total = $query[1];
         return $total;
     }
 
@@ -21,16 +21,16 @@ class MessageRepository extends EntityRepository
      * 取得每頁資料
      * @param int $offset 目前第幾頁
      * @param int $pageLimit 每頁最大數量資料
-     *
+     * @return array $query select 搜尋完的結果
      */
     public function getPages($offset, $pageLimit)
     {
         $dql = "SELECT m.sn, m.name, m.time, m.msg FROM Message m";
         $query = $this->_em
-                    ->createQuery($dql)
-                    ->setFirstResult($offset)
-                    ->setMaxResults($pageLimit)
-                    ->getScalarResult();
+            ->createQuery($dql)
+            ->setFirstResult($offset)
+            ->setMaxResults($pageLimit)
+            ->getScalarResult();
         return $query;
     }
